test(threads): 拆分 spec root 消息测试
